requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref
async-timeout==4.0.2
attrs==22.2.0
backoff==1.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django
django-allauth==0.52.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt==5.2.2
emoji==1.7.0
frozenlist==1.3.3
httpie==3.2.1
idna==3.4
joblib==1.2.0
markdown-it-py==2.1.0
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
oauthlib==3.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz
redis==4.4.2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.2.0
six==1.16.0
spotipy==2.22.0
sqlparse
text2emotion==0.0.5
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
google-cloud-vision
openai
cv2